requirements.txt
numpy~=2.2.6
pandas~=2.2.3
selenium~=4.33.0
webdriver-manager~=4.0.2
requests~=2.32.3
beautifulsoup4~=4.13.4
undetected-chromedriver~=3.5.5
Crawl4AI~=0.6.3
transformers~=4.53.0
playwright~=1.53.0
protobuf~=5.29.5
extruct~=0.18.0
w3lib~=2.3.1
python-dateutil~=2.9.0.post0
datefinder~=0.7.3
PyMuPDF
bs4~=0.0.2
openai~=1.95.1